getFortResults
Added a function to easily obtain stats for a given fort.
Usage: Function that takes one argument, number of users whose stats it has to calculate (starts from the top). Default Argument value is 50
Pulls username and calculates stats for the given fort for that user and logs it in the console.
Takes 10 seconds per user